revert to using floats

+++ src/setup.hpp
@@ -3,6 +3,6 @@
 
 namespace setup 
 {
-  using real_t = double;
+  using real_t = float;
   using arr_1D_t = blitz::Array<setup::real_t, 1>;
 };
